Use more decomposition, string_view, and implicit sizes in ESM code

+++ components/esm/luascripts.cpp
@@ -38,7 +38,7 @@ std::string ESM::loadLuaBinaryData(ESMReader& esm)
     {
         esm.getSubHeader();
         data.resize(esm.getSubSize());
-        esm.getExact(data.data(), static_cast<int>(data.size()));
+        esm.getExact(data.data(), data.size());
     }
     return data;
 }
